Z_AI_Medbot/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
pillow
pytesseract
pdf2image
python-dotenv
requests
openai
pytest
pytest-asyncio  